new ExamServlet. moved dao mocks from test/dao to dao. new methods in QuestionCollection and ChoiceCollection. New Interfaces for them

+++ src/main/java/model/Chapter.java
@@ -4,9 +4,9 @@ public enum Chapter{
 	THE_FUNDAMENTALS_OF_TESTING(1), LIFE_CYCLES(2), STATIC_TESTING(3), TEST_DESIGN_TECHNIQUES(4),
 	TEST_MANAGEMENT(5),TOOL_SUPPORT_FOR_TESTING(6), THE_EXAMINATION (7);
 	
-    private final int no;
+    private final long no;
     
-    private Chapter(int no) {
+    private Chapter(long no) {
         this.no = no;
 }
 
@@ -14,12 +14,14 @@ public enum Chapter{
     		"TEST DESIGN TECHNIQUES", "TEST MANAGEMENT", "TOOL SUPPORT FOR TESTING", "THE EXAMINATION"};
 
     
-    public int getChapterNo() {
+    public long getChapterNo() {
         return this.no;
     }
     
+
+    
     public String getHeader(){
-    	return this.heading[this.no + 1];
+    	return this.heading[(int) (this.no + 1)];
     }
   
 }
